+++ Craftsman/Builders/Endpoints/ControllerBuilder.cs
@@ -16,17 +16,17 @@ public class ControllerBuilder
     public void CreateController(string solutionDirectory, string srcDirectory, string entityName, string entityPlural, string projectBaseName, bool isProtected)
     {
         var classPath = ClassPathHelper.ControllerClassPath(srcDirectory, $"{FileNames.GetControllerName(entityPlural)}.cs", projectBaseName, "v1");
-        var fileText = GetControllerFileText(classPath.ClassNamespace, entityName, entityPlural, solutionDirectory, srcDirectory, projectBaseName, isProtected);
+        var fileText = GetControllerFileText(classPath.ClassNamespace, entityPlural, solutionDirectory, srcDirectory, projectBaseName, isProtected);
         _utilities.CreateFile(classPath, fileText);
     }
 
-    public static string GetControllerFileText(string classNamespace, string entityName, string entityPlural, string solutionDirectory, string srcDirectory, string projectBaseName, bool usesJwtAuth)
+    public static string GetControllerFileText(string classNamespace, string entityPlural, string solutionDirectory, string srcDirectory, string projectBaseName, bool usesJwtAuth)
     {
         // TODO create an attribute factory that can order them how i want and work more dynamically
 
         var endpointBase = FileNames.EndpointBaseGenerator(entityPlural);
 
-        var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", entityName, projectBaseName);
+        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", entityPlural, projectBaseName);
         var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
         var featureClassPath = ClassPathHelper.FeaturesClassPath(srcDirectory, "", entityPlural, projectBaseName);
         var permissionsClassPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, "", projectBaseName);