+++ Craftsman/Builders/SolutionBuilder.cs
@@ -54,7 +54,7 @@ public class SolutionBuilder
         _fileSystem.Directory.CreateDirectory(ClassPathHelper.WebApiMiddlewareClassPath(srcDirectory, "", projectBaseName).ClassDirectory);
 
         // additional from what was other projects
-        _fileSystem.Directory.CreateDirectory(ClassPathHelper.DtoClassPath(solutionDirectory, "", "", projectBaseName).ClassDirectory);
+        _fileSystem.Directory.CreateDirectory(ClassPathHelper.DtoClassPath(srcDirectory, "", "", projectBaseName).ClassDirectory);
         _fileSystem.Directory.CreateDirectory(ClassPathHelper.ExceptionsClassPath(solutionDirectory, "").ClassDirectory);
         _fileSystem.Directory.CreateDirectory(ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName).ClassDirectory);
         _fileSystem.Directory.CreateDirectory(ClassPathHelper.SharedDtoClassPath(solutionDirectory, "").ClassDirectory);