+++ Craftsman/Builders/Tests/IntegrationTests/PatchCommandTestBuilder.cs
@@ -31,7 +31,7 @@ public class PatchCommandTestBuilder
         var testUtilClassPath = ClassPathHelper.IntegrationTestUtilitiesClassPath(testDirectory, projectBaseName, "");
         var fakerClassPath = ClassPathHelper.TestFakesClassPath(testDirectory, "", entity.Name, projectBaseName);
         var exceptionClassPath = ClassPathHelper.ExceptionsClassPath(testDirectory, "");
-        var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", entity.Name, projectBaseName);
+        var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", entity.Plural, projectBaseName);
         var featuresClassPath = ClassPathHelper.FeaturesClassPath(srcDirectory, featureName, entity.Plural, projectBaseName);
 
         var myProp = entity.Properties.FirstOrDefault(e => e.Type == "string" && e.CanManipulate);