+++ CHANGELOG.md
@@ -76,6 +76,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Add `FakeItEasy` and `FakeItEasyAnalyzer` to unit test project
 
+* Colocate DTOs with entities by default. Can be moved to shared kernel when needed
+
 * Default db provider set to postgres
 
 * Removed `?`'s on strings in  `BaseEntity`